refactor #205: rename position card
