+++ components/templates/tournamentDetail/components/tournamentSummary/components/column/index.module.css
