+++ components/templates/tournamentDetail/components/tournamentSummary/components/column/index.tsx
