+++ components/templates/tournamentDetail/components/tournamentSummary/index.module.css
