--- components/templates/tournaments/components/rowCards/index.module.css
@@ -1,65 +1,0 @@
-.title {
-  color: var(--text-color-primary);
-  margin-bottom: 0;
-  margin-top: 0;
-}
-
-.podium {
-  flex: 0 0 calc(50% - 8px);
-  width: calc(50% - 8px);
-}
-
-.podium:first-child {
-  flex: 0 0 100%;
-  width: 100%;
-}
-
-.card {
-  align-items: center;
-  background-color: var(--background-quaternary);
-  border-radius: 5px;
-  display: flex;
-  flex: 1;
-  flex-direction: row;
-  flex-wrap: wrap;
-  gap: 16px;
-  justify-content: space-between;
-  padding: 16px;
-  width: 100%;
-}
-
-.container {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  gap: 16px;
-  margin-top: 16px;
-  width: 100%;
-}
-
-.columnsContainer {
-  display: flex;
-  flex: 1;
-  flex-direction: row;
-  flex-wrap: wrap;
-}
-
-@media (max-width: 480px) {
-  .podium {
-    flex: 0 0 100%;
-    width: 100%;
-  }
-
-  .card {
-    align-items: flex-start;
-    flex-direction: column;
-  }
-
-  .columnsContainer {
-    width: auto;
-  }
-
-  .columnsContainer > *:not(:first-child) {
-    display: none;
-  }
-}
