+++ components/templates/tournaments/components/tournamentSummary/components/column/index.module.css
