+++ components/templates/tournaments/components/tournamentSummary/components/column/index.tsx
