+++ components/templates/tournaments/components/tournamentSummary/index.module.css
@@ -1,0 +1,31 @@
+.podium {
+  flex: 0 0 calc(50% - 8px);
+  width: calc(50% - 8px);
+}
+
+.podium:first-child {
+  flex: 0 0 100%;
+  width: 100%;
+}
+
+.columnsContainer {
+  display: flex;
+  flex: 1;
+  flex-direction: row;
+  flex-wrap: wrap;
+}
+
+@media (max-width: 480px) {
+  .podium {
+    flex: 0 0 100%;
+    width: 100%;
+  }
+
+  .columnsContainer {
+    width: auto;
+  }
+
+  .columnsContainer > *:not(:first-child) {
+    display: none;
+  }
+}
